+++ src/fruit/health/client/entities/PlanData.java
@@ -4,7 +4,7 @@ public class PlanData
 {
     public String planName;
     public Integer premium;
-    public Integer copay; // %age
+    public Double copay; // %age
     public Integer deductible;
     public Integer oopMax;
